+++ contracts/safex/libraries/SAFEXLinkedList.sol
@@ -23,44 +23,32 @@ library SAFEXLinkedList {
         self.lmp = lmp_;
     }
 
-    function _heads(
-        PriceLinkedList storage self
-    ) internal view returns (uint256, uint256) {
+    function _heads(PriceLinkedList storage self) internal view returns (uint256, uint256) {
         return (self.askHead, self.bidHead);
     }
 
-    function _askHead(
-        PriceLinkedList storage self
-    ) internal view returns (uint256) {
+    function _askHead(PriceLinkedList storage self) internal view returns (uint256) {
         return self.askHead;
     }
 
-    function _bidHead(
-        PriceLinkedList storage self
-    ) internal view returns (uint256) {
+    function _bidHead(PriceLinkedList storage self) internal view returns (uint256) {
         return self.bidHead;
     }
 
-    function _mktPrice(
-        PriceLinkedList storage self
-    ) internal view returns (uint256) {
+    function _mktPrice(PriceLinkedList storage self) internal view returns (uint256) {
         if (self.lmp == 0) {
             revert NoMatchPrice(self.askHead, self.bidHead, self.lmp);
         }
-        if(self.lmp < self.bidHead) {
+        if (self.lmp < self.bidHead) {
             return self.bidHead;
-        } else if(self.lmp > self.askHead && self.askHead != 0) {
+        } else if (self.lmp > self.askHead && self.askHead != 0) {
             return self.askHead;
         } else {
             return self.lmp;
         }
     }
 
-    function _next(
-        PriceLinkedList storage self,
-        bool isBid,
-        uint256 price
-    ) internal view returns (uint256) {
+    function _next(PriceLinkedList storage self, bool isBid, uint256 price) internal view returns (uint256) {
         if (isBid) {
             return self.bidPrices[price];
         } else {
@@ -69,11 +57,7 @@ library SAFEXLinkedList {
     }
 
     // for bidPrices, lower ones are next, for askPrices, higher ones are next
-    function _insert(
-        PriceLinkedList storage self,
-        bool isBid,
-        uint256 price
-    ) internal {
+    function _insert(PriceLinkedList storage self, bool isBid, uint256 price) internal {
         if (isBid) {
             uint256 last = 0;
             uint256 head = self.bidHead;
@@ -91,7 +75,7 @@ library SAFEXLinkedList {
                     head = self.bidPrices[head];
                     last = next;
                 } else if (price > next) {
-                    if(next == 0) {
+                    if (next == 0) {
                         // Insert price at the end of the list
                         self.bidPrices[head] = price;
                         self.bidPrices[price] = 0;
@@ -147,11 +131,7 @@ library SAFEXLinkedList {
     }
 
     // show n prices shown in the orderbook
-    function _getPrices(
-        PriceLinkedList storage self,
-        bool isBid,
-        uint n
-    ) internal view returns (uint256[] memory) {
+    function _getPrices(PriceLinkedList storage self, bool isBid, uint256 n) internal view returns (uint256[] memory) {
         uint256 i = 0;
         uint256[] memory prices = new uint256[](n);
         for (
